refactor: extract command implementations into separate modules
